JK/pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "pi5-ai-assistant"
version = "0.1.0"
description = "Offline Multi-Modal AI Assistant for Raspberry Pi 5"
readme = "README.md"
requires-python = ">=3.9"
license = {text = "MIT"}
authors = [
  {name = "Suhas", email = "suhas@example.com"}
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
dependencies = [
    "luma.oled",
    "luma.core",
    "RPi.GPIO",
    "chromadb",
    "llama-cpp-python",
    "SpeechRecognition",
    "gTTS",
    "pyaudio",
    "ultralytics",
    "numpy",
    "pillow",
    "sounddevice"
]

[project.urls]
"Homepage" = "https://github.com/example/pi5-ai-assistant"

[tool.setuptools.packages.find]
where = ["."]
include = ["app*"]
